+++ src/components/AppViews/PerspectiveAppview.vue
@@ -22,42 +22,11 @@
     <p class='description txt'>
       {{ pers.description }}
     </p>
-    <div class='tags-wrapper'>
-      <div class='tags'>
-        <transition-group name='fade'>
-          <view-tag v-for='tag in getSmartLabels'
-            :key='tag.name'
-            :name='tag.name'
-            :fixed='tag.fixed'
-            icon='tag'
-            back-color='#83B7E2'
-          />
-        </transition-group>
-      </div>
-      <transition name='fade'>
-        <div class='tags' v-if="search && search !== ''">
-          <view-tag
-            icon='search'
-            back-color='#88DDB7'
-            :name='search'
-            :fixed='false'
-            @click="v => search = ''"
-          />
-        </div>
-      </transition>
-      <div class='tags'>
-        <transition-group name='fade'>
-          <view-tag v-for='tag in labels'
-            :key='tag.name'
-            :name='tag.name'
-            :fixed='false'
-            icon='tag'
-            @click='deselect'
-            back-color='#FF6B66'
-          />
-        </transition-group>
-      </div>
-    </div>
+    <view-tags
+      :is-default-perspective='isDefaultPerspective'
+      :search='search'
+      @clearsearch="v => search = ''"
+    />
   </div>
 </template>
 
@@ -67,8 +36,8 @@ import { Component, Vue, Watch } from 'vue-property-decorator'
 import { State, Getter, namespace } from 'vuex-class'
 
 import DynamicFontawesome from '@/components/DynamicFontawesome.vue'
-import Tag from '@/components/AppViews/AppviewComponents/AppviewIcon.vue'
 import DropdownFinder from '@/components/AppViews/AppviewComponents/DropdownFinder.vue'
+import AppviewTags from '@/components/AppViews/AppviewComponents/AppviewTags.vue'
 
 import { SmartPerspective, Label } from '../../interfaces/app'
 
@@ -82,45 +51,28 @@ library.add(faSearch)
 @Component({
   components: {
     'dynamic-ft-icon': DynamicFontawesome,
-    'view-tag': Tag,
     'drop-finder': DropdownFinder,
+    'view-tags': AppviewTags,
   },
 })
 export default class PerspectiveAppview extends Vue {
   @State('perspectiveData') pers!: SmartPerspective
   @Getter isDesktop!: boolean
+  @Getter isDefaultPerspective!: boolean
+
+  search: string = ''
 
   @labelVuex.Getter sortedLabels!: Label[]
   @labelVuex.Getter getLabelsByIds!: (ids: string[]) => Label[]
-
-  labels: Label[] = []
-  search: string = ''
-
-  select(lab: Label) {
-    if (!this.labels.find(el => el.id === lab.id))
-      this.labels.push(lab)
-  }
-  deselect({name, icon}: {name: string, icon: string}) {
-    if (icon === 'tag') {
-      const index = this.labels.findIndex(el => el.name === name)
-      this.labels.splice(index, 1)
-    }
-  }
-
-  get getSmartLabels(): Label[] {
-    const labels: any[] = []
-    if (this.pers.name === 'Inbox')
-      labels.push({
-        name: 'Inbox',
-        fixed: true,
-      })
-    return labels
-  }
 }
 
 </script>
 
 <style scoped>
+
+.description {
+  margin: 30px 0;
+}
 
 .component {
   padding-top: 20px;
@@ -135,22 +87,6 @@ export default class PerspectiveAppview extends Vue {
   top: 50%;
   right: 0;
   transform: translateY(-50%);
-}
-
-.description {
-  margin: 30px 0;
-}
-
-.tags-wrapper {
-  margin-bottom: 30px;
-}
-
-.tags + .tags {
-  margin-top: 6px;
-}
-
-.tag + .tag {
-  margin-left: 4px;
 }
 
 .title {
